Add unit tests for GlobalState::update_global_state
These unit tests have been added in order to check that the modification of the
`GlobalState` object is intended with the invocation of `update_global_state`.
Specific unit tests have also been added to verify the behavior of ovewriting
previously stored values of hashmap, or not updating the stored value of the LRU
in order to verify that these behaviors are explicitly intended.