Add unit tests for GlobalState::remove_handles
The `remove_handles` method of `GlobalState` ends up modifying the stored data in
`GlobalState` in somewhat interesting ways.  In order to ensure that this behavior is
correct, and to draw attention to its implementation, unit tests have been added in
order to verify and assert that the current behavior is the intended behavior.