Add unit tests for `handle_received_txns`
The function `handle_received_txns` is currently written in such a way that makes it
difficult to write unit tests for.  In order to remedy this, a new error enum has been
created called `HandleReceivedTxnsError` which codifies the different failure reasons
that can cause the transactions not to be broadcast to the given receiver.  Additionally,
since the original return value is a Vec, and in order to prevent a double Vec allocation
unnecessarily, this has been implemented as an Iterator to make it collectable with a
single allocation.

The original function `handle_received_txns` has modified to create this `Iterator`
called `HandleReceivedTxns` and to collect it's results into a `Vec` with its errors
being mapped back to a `BuildError`.

With this modification unit tests can be added to target the `HandleReceivedTxns`
`Iterator` implementation directly in order to assert the failure and success cases
of the function's processing.