Add unit tests for `connect_to_events_service`
In it's current state the function `connect_to_events_service` is not easily testable with reproducible /
controlled conditions.  Due to this, and the nature of what the function `connect_to_events_service`
is meant to be doing, the code has been refactored to introduce a `trait`, `HotShotEventsService`,
that can be interacted with that exposes the relevant API endpoints analogs.  This `trait` can then be
implemented with different implementations that makes it suitable for testing with mocks.

An implementation of the trait `HotShotEventsService` has also been provided called
`HotShotEventsServiceTideDiscoClient` that implements the `trait` with a tide disco client.  This
is the path that allows for the original behavior to be maintained.

Additionally `connect_to_events_service` has been refactored to return a `Result` instead of an
`Option` so that a new enum type, `ConnectToEventsServiceError`, can be returned when the
specific failure condition arises.

The function `connect_to_events_service` now takes a `HotShotEventsService` client in order
to perform the various function calls.

All of these changes allow for `connect_to_events_service` to be effectively tested for its edge
cases with specifically designed mock implementations that will trigger the specific conditions
without the need to actually establish a connection to a static external source.