Replace version and edition shorthands in crates
The `cargo-sort` utility seems to not be able to understand the shorthand assignment
for the workspace.

Example:

This is problematic and causes cargo-sort to error.

```
version.workspace = true
```

This works with cargo-sort
```
version = { workspace = true }
```

This replaces the shorthand with the longer form so that `cargo-sort` can
work and not interfere with the nix precommit hook.