Refactor order tests for deterministic behavior
The current order tests can be flaky at times as they rely on waiting for the builder's
logic to process via `async_sleep` without being able to determine if its in a state that
is ready.

To address this specifically the code has been refactored to work with `ProxyGlobalState`
instead of the `GlobalState` directly as `ProxyGlobalState` already has some mechanisms
for waiting for things to be completed.  This also aids in simplifying a bit of the internal
logic, as we were reimplementing what `ProxyGlobalState` was already doing.

In addition to these changes this also attempts to refactor the code in such a way as to
remove the conditionals that occur within the body of the for each loops.  This is done
by adding an enum for round behavior that can adjust the transactions returned by the
bundle call as needed.

The order of progression within the for each loop has also been adjusted to perform the
consensus portion at the top of the loop before transactions are submitted.  Since this
order isn't that important, it seems better to have the consensus to occur at the top of
the loop rather than after the transactions as it makes it more clear which portion of the
logic we are currently in.  I.E. are we currently considering consensus' role, or the
builder's role.