Fix BuilderState unecessary forking
The current `BuilderState` extending behavior fails to anticipate enough of the
various edge cases that determine whether the current `BuilderState` is the best,
or one of the only `BuilderState`s that should be extended.  Due to this behavior
many more `BuilderState`s are spawned that really should not be spawned.  This
causes a lot of unecessary computation and memory usage that will never get cleaned
up as the `Sender`s that are registered for them get overwritten by subsequent
clones that have the same `BuilderStateId`. This effectively causes an async
memory leak, async processing leak, and a potential fork-bomb that will effectively
prevent the builder from being responsive.

Additionally, the race condition that stems from multiple `BuilderState`s spawning
a clone that extends them for the next `QuorumProposal` means that we often
recieve an unexpected list of transactions as a result.  This behavior can be
best seen by the test
marketplace::testing::order_test::test_builder_order_chain_fork`.

To fix these issues a few changes have been maded.  First a new method has been
added called `spawn_clone_that_extends_self` that ensures a standard way of
spawning a new `BuilderState` that extends the current `BuilderState`. Another
method has been added called `am_i_the_best_builder_state_to_extend` that
attempts to ensure that we are the best fit, or one of the only fits that can
extend the current `BuilderState` based on the given `QuorumProposal`. Finally,
as a safe guard, a check has been added when calling `spawn_clone` that ensures
that the same `BuilderStateId` will not be registered more than once.  IE, if
there is already an async process spawned, there is no need to spawn another.

Remove newly added asserts as they fail to represent the correct anticipated intermediate state