Fix best `BuilderState` algorithm
With the addition of the the `BuiltForProposedBlock` to the `spawned_builder_states`
we can not make better decisions about which `BlockerState` a given `QuorumProposal`
can extend.  As such, this splits the determiniation of the best `BuilderState`s to
extend from the `am_i_the_best_builder_state_to_extend` call.

Additionally the new `best_builder_states_to_extend` function is deliberately
made into a function rather than a `method` in order to ensure that information
contained within the current `BuilderState` that is only visible to that
`BuilderState` is used.  This should allow for more consistent determiniations of
the proper `BuilderState` to extend.