Reorder `BuilderState` transaction members
This change groupst he transaction structures in `BuilderState` together for better
association. For instance `tx_queue` should always be kept in sync with `txns_in_queue`
but with these parameters being separated in the structure definition it seems to imply
that they might have a looser association.

Additionally, this renames `txns_in_queue` to `txn_commits_in_queue` to help add
clarity for what is being stored, and why.

Added some comments on these fields to help expand on what they are for, and why they
are grouped together.