Refactor `process_block_request` to be "Happy Path"
The method `process_block_request` currently has a few nested conditionals that evaluate
two branches.  Each "negative" case of the branch is accompanied by a return without
further progress.  These nested statements have been flattened with their "negative"
case being placed higher.  This should allow for less cognitive load to be needed when
reading the code and understanding what it does in the future.

Additionally a comment has been added to the method to explain what this function is
doing and how it returns its response.