Refactor `event_loop` into individual functions for each channel
The main purpose of `event_loop` is to process requests as they come in through their
channels as soon as possible.  However, it is incredibly easy to get lost inside of the
`event_loop` itself as it has many different control flows happening in a large function
body.

This has been refactored so that each receiving channel decision has been split into
a separate function whose conditional paths have been converted into a "Happy Path".
Each helper method in addition to the `event_loop` itself has had a comment added to
it to help describe what it is doing, and in an effort to add clarity for the
functionality that is occuring.