Refactor `build_block` to take the "Happy Path"
The method `build_block` has a single conditional whose else statement just logs and
returns a `None`. This method has been refactored to move that result to the top of
method for an early return.  This allows the remaining "Happy Path" to be flattened
and live outside of the conditional for better clarity of purpose.

Additionally a comment has been added to `build_block` to explain its function and
the steps taken to produce the result.