Add `TEST_CHANNEL_BUFFER_SIZE` for broadcast channel size
Based on discussion with @dailinsubjam:
https://github.com/EspressoSystems/marketplace-builder-core/pull/152#discussion_r1790803638

Currently all of the `broadcast` calls are being created with a hard-coded buffer size of
`32`.  This decision is somewhat arbitrary, but for clarity it is beneficial to replace
these hard-coded values with a `const` value that can have a describing comment.

The variable `TEST_CHANNEL_BUFFER_SIZE` represents this value with a comment that is
provided to add context.