Replace parameter with hard-coded constant
Based on discussion with @shenkeyao:
https://github.com/EspressoSystems/marketplace-builder-core/pull/152#discussion_r1792440655

We don't expect the number of subsequent blocks that are needed in order to reach
finalization to have a loose or changing requirement.  As such, it makes more sense to
keep the period as a hard-coded constant rather than a configurable value.

This change removes the `allow_empty_block_period` parameter, and adds the
`ALLOW_EMPTY_BLOCK_PERIOD` constant instead.